[Common] Fix Sonar reliability gate and reduce open issues
Fix integer overflow in FlinkSavepointServiceImpl pagination, harden
regex patterns across common/flink modules, annotate deprecated APIs,
and expand SonarCloud ignore rules for migration-specific code smells.

+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/service/impl/FlinkSavepointServiceImpl.java
@@ -471,7 +471,7 @@ public class FlinkSavepointServiceImpl extends ServiceImpl<FlinkSavepointMapper,
             .eq(FlinkSavepoint::getType, CHECKPOINT.get())
             .orderByDesc(FlinkSavepoint::getTriggerTime)
             .page(
-                new Page<>(1, cpThreshold + 1));
+                new Page<>(1, (long) cpThreshold + 1L));
 
         if (CollectionUtils.isEmpty(savepointPage.getRecords())
             || savepointPage.getRecords().size() <= cpThreshold) {
